backend/emotion_detection/requirements.txt
# Emotion Detection Requirements
# Based on face_and_emotion_detection repository

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Face Detection and Recognition
face-recognition>=1.3.0
dlib>=19.24.0

# Image Processing
Pillow>=10.0.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
scipy>=1.11.0

# Model Training (optional, for retraining)
scikit-learn>=1.3.0
matplotlib>=3.7.0
